backend/pyproject.toml
[project]
name = "ai-radar-dxc-backend"
version = "0.1.0"
description = "AI Radar DXC — Backend API for AI Use Case Scoring & Prioritization"
requires-python = ">=3.11"
dependencies = [
    # ─── Web Framework ──────────────────────────────────────
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # ─── Database ───────────────────────────────────────────
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # ─── Vector Store ───────────────────────────────────────
    "qdrant-client>=1.7.0",

    # ─── AI / RAG ───────────────────────────────────────────
    "llama-index-core>=0.10.0",
    "llama-index-vector-stores-qdrant>=0.1.0",
    "llama-index-embeddings-openai>=0.1.0",

    # ─── LLM SDKs ──────────────────────────────────────────
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-genai>=0.8.0",
    "mistralai>=1.0.0",

    # ─── Embeddings ─────────────────────────────────────────
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",

    # ─── Scraping ───────────────────────────────────────────
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.12.0",

    # ─── Cache ──────────────────────────────────────────────
    "redis>=5.0.0",

    # ─── PDF Export ─────────────────────────────────────────
    "weasyprint>=60.0",
    "jinja2>=3.1.0",

    # ─── Utilities ──────────────────────────────────────────
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.27.0",
    "ruff>=0.2.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
